customer_churn_ml/pyproject.toml
[project]
name = "mall-customers-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.6",
    "numpy>=2.3.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pyarrow>=21.0.0",
    "pytest>=8.4.2",
    "pyyaml>=6.0.2",
    "ruff>=0.12.12",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
]
